refactor(RHOAIENG-79876): split Prometheus types into separate k8s-core subpath
Address reviewer feedback on PR #8994:
- Split Prometheus types from TrustyAI into ./prometheus subpath
- Move PrometheusQueryResponse to k8s-core/prometheus
- Consolidate duplicate imports from same module
- Add prometheusK8sCoreExports.spec.ts for subpath verification
