+++ frontend/src/__mocks__/mockPrometheusServing.ts
@@ -1,5 +1,7 @@
-import type { PrometheusQueryRangeResponseDataResult } from '@odh-dashboard/k8s-core/trustyai';
-import { PrometheusQueryRangeResponse } from '#~/types';
+import type {
+  PrometheusQueryRangeResponse,
+  PrometheusQueryRangeResponseDataResult,
+} from '@odh-dashboard/k8s-core/prometheus';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
