+++ frontend/src/api/prometheus/pvcs.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { PersistentVolumeClaimKind } from '@odh-dashboard/k8s-core';
-import { PrometheusQueryResponse } from '#~/types';
+import type { PrometheusQueryResponse } from '@odh-dashboard/k8s-core/prometheus';
 import { POLL_INTERVAL } from '#~/utilities/const';
 import usePrometheusQuery from './usePrometheusQuery';
