+++ frontend/src/api/prometheus/useQueryRangeResourceData.ts
@@ -2,7 +2,7 @@ import React from 'react';
 import { FetchOptions } from '@odh-dashboard/ui-core/hooks/useFetchState';
 import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import { TimeframeStep, TimeframeTimeRange } from '@odh-dashboard/ui-core/utilities/metrics';
-import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/k8s-core/trustyai';
+import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/k8s-core/prometheus';
 import type { PendingContextResourceData } from '#~/types';
 import { useMakeFetchObject } from '#~/utilities/useMakeFetchObject';
 import usePrometheusQueryRange, { ResponsePredicate } from './usePrometheusQueryRange';
