+++ packages/gpuaas/src/hooks/useBorrowingLendingMetrics.ts
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import usePrometheusQueryRange from '@odh-dashboard/internal/api/prometheus/usePrometheusQueryRange';
 import { ClusterQueueKind } from '@odh-dashboard/k8s-core';
-import type { PrometheusQueryRangeResponseData } from '@odh-dashboard/internal/types';
 import type {
+  PrometheusQueryRangeResponseData,
   PrometheusQueryRangeResponseDataResult,
   PrometheusQueryRangeResultValue,
-} from '@odh-dashboard/k8s-core/trustyai';
+} from '@odh-dashboard/k8s-core/prometheus';
 import {
   ACCELERATOR_RESOURCE_PREFIXES,
   ACCELERATOR_RESOURCE_REGEX,
