+++ packages/k8s-core/package.json
@@ -5,6 +5,7 @@
   "version": "0.0.0",
   "exports": {
     ".": "./src/index.ts",
+    "./prometheus": "./src/prometheusTypes.ts",
     "./trustyai": "./src/trustyaiTypes.ts"
   },
   "scripts": {
