+++ packages/model-serving/src/components/metrics/bias/BiasMetricConfigSelector.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import { useModelBiasData } from '@odh-dashboard/internal/concepts/trustyai/context/useModelBiasData';
-import type { BiasMetricConfig } from '@odh-dashboard/k8s-core/trustyai';
-import { BiasMetricType, TrustyInstallState } from '@odh-dashboard/k8s-core/trustyai';
+import {
+  BiasMetricType,
+  TrustyInstallState,
+  type BiasMetricConfig,
+} from '@odh-dashboard/k8s-core/trustyai';
 import {
   MultiSelection,
   SelectionOptions,
