+++ packages/model-serving/src/components/metrics/utils.tsx
@@ -7,12 +7,12 @@ import { isEnumMember } from '@odh-dashboard/foundation';
 import { TimeframeTitle, BreadcrumbItemType } from '@odh-dashboard/ui-core/types/metrics';
 import { PROMETHEUS_REQUEST_RESOLUTION } from '@odh-dashboard/ui-core/utilities/metrics';
 import type {
-  PrometheusQueryRangeResultValue,
   BaseMetricRequest,
   BaseMetricRequestInput,
   BiasMetricConfig,
 } from '@odh-dashboard/k8s-core/trustyai';
 import { BiasMetricType } from '@odh-dashboard/k8s-core/trustyai';
+import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/k8s-core/prometheus';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import { BIAS_CHART_CONFIGS, BIAS_THRESHOLD_COLOR } from './const';
 import {
